[PDDF] Add FAN's get_fault(), fix int overflow in PSU, rm unused var (#24051)
* [PDDF] Implement get_fault() in PddfFan

* [PDDF] Remove unused variable in struct XCVR_ATTR

* [PDDF] Fix LINEAR11 int overflow in PSU

+++ platform/pddf/i2c/modules/psu/driver/pddf_psu_api.c
@@ -193,7 +193,7 @@ static u8 psu_get_vout_mode(struct i2c_client *client)
 static long pmbus_linear11_to_int(u16 value, int multiplier)
 {
     s16 exponent;
-    s32 mantissa;
+    s64 mantissa;
 
     exponent = two_complement_to_int(value >> 11, 5, 0x1f);
     mantissa = two_complement_to_int(value & 0x7ff, 11, 0x7ff);
